fix(threading): completely revert the use of virtual threads
Even when running in JRE 25, some users of the SDK experience app freezes when virtual threads are used. In those instances, responses sent by the runtime over websocket don't seem to be picked up by the new ThreadPerTask executor. This needs further investigation but for now, roll back the use of virtual threads.